chore: reorganize files in common modules
