+++ packages/backend/src/modules/artworks/enums/file-type.enum.ts
