+++ packages/backend/src/modules/artworks/exceptions/upload-image.exception.ts
