+++ packages/backend/src/modules/auth/exceptions/github-auth.exception.ts
