+++ packages/backend/src/modules/auth/exceptions/token.exception.ts
