+++ packages/backend/test/modules/auth/guards/cookie.auth.guard.spec.ts
@@ -1,10 +1,10 @@
 import { ExecutionContext, HttpStatus } from '@nestjs/common';
 
+import { AuthService } from '@/src/modules/auth/auth.service';
 import {
   TokenErrorCode,
   TokenException,
-} from '@/src/common/exceptions/auth/token.exception';
-import { AuthService } from '@/src/modules/auth/auth.service';
+} from '@/src/modules/auth/exceptions/token.exception';
 import { CookieAuthGuard } from '@/src/modules/auth/guards/cookie.auth.guard';
 import { createTestingModuleWithoutDB } from '@/test/utils/module-builder.util';
